% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.package-package.R
\docType{package}
\name{example.package-package}
\alias{example.package}
\alias{example.package-package}
\title{example.package: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/significantstats/example.package}
  \item Report bugs at \url{https://github.com/significantstats/example.package/issues}
}

}
\author{
\strong{Maintainer}: Kevin Brosnan \email{kevin.c.brosnan@gmail.com} (\href{https://orcid.org/0000-0002-0952-2134}{ORCID})

}
\keyword{internal}
