% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.R
\name{f_to_c}
\alias{f_to_c}
\title{Transform fahrenheit to celsius}
\usage{
f_to_c(temp_f)
}
\arguments{
\item{temp_f}{numeric vector (vs scalar) of temperatures in fahrenheit}
}
\value{
numeric vector of temperatures in celsius
}
\description{
Transform fahrenheit to celsius
}
\examples{
f_to_c(c(0, 32, 100))
}
