% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fit.R
\name{fit_lm_int}
\alias{fit_lm_int}
\title{Title}
\usage{
fit_lm_int(lm_data)
}
\arguments{
\item{lm_data}{simulated data from lm_data}
}
\value{
lm model
}
\description{
Title
}
\examples{
df <- lm_data(10)
fit_lm_int(df)
}
