% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello(x, y = "!")
}
\arguments{
\item{x}{a string}

\item{y}{a string that will be the puctuation of the phrase, default value is !}
}
\description{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\details{
blah blah
}
