% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_data.R
\name{lm_data}
\alias{lm_data}
\title{Title}
\usage{
lm_data(n = 10)
}
\arguments{
\item{n}{Sample size for simulated lm data}
}
\value{
tibble, n rows and 3 columns: y, x_1, x_2
}
\description{
Title
}
\examples{
lm_data(10)
}
