% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_example}
\alias{lm_example}
\title{lm_example}
\format{
100 observations in 3 variables: y, x_1, x_2
\describe{
  \item{x_1}{continuous}
  \item{x_2}{binary}
}
}
\source{
google.com
}
\usage{
lm_example
}
\description{
Simulated example lm data
}
\keyword{datasets}
