% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_matrices.R
\name{multiply}
\alias{multiply}
\title{Multiply two matrices}
\usage{
multiply(A, B)
}
\arguments{
\item{A}{a matrix.}

\item{B}{a matrix.}
}
\value{
The product of \code{A} and \code{B}.
}
\description{
Multiply two matrices
}
\examples{
X <- matrix(1:6,2,3)
Y <- matrix(1:15,3,5)
multiply(X, Y)
}
