% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Title}
\usage{
plot_fit(lm.mod, lm_data, new_data, interaction = FALSE)
}
\arguments{
\item{lm.mod}{lm object}

\item{lm_data}{n x 3 tibble, output from lm_data}

\item{new_data}{new data.frame or tibble with columns y, x_1, x_2}

\item{interaction}{logical, changes plot title according to fit}
}
\value{
ggplot
}
\description{
Title
}
\examples{
data <- lm_data(10)
plot_fit(lm_fit(data), data, data, interaction = FALSE)
}
