% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mean_sd.R
\name{sim_mean_sd}
\alias{sim_mean_sd}
\title{sim_mean_sd}
\usage{
sim_mean_sd(n, mu = 2, sigma = 3)
}
\arguments{
\item{n}{the sample size}

\item{mu}{the true mean}

\item{sigma}{the true SD}
}
\value{
a tibble with sample mean and SD
}
\description{
This function produces a sample from a normal distribution and returns the sample mean and sample SD.
}
\examples{

sim_mean_sd(30)

}
