% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_regression.R
\name{sim_regression}
\alias{sim_regression}
\title{Simulate from an SLR}
\usage{
sim_regression(n, beta0 = 2, beta1 = 3)
}
\arguments{
\item{n}{sample size}

\item{beta0}{intercept}

\item{beta1}{slope}
}
\value{
tbl_df with one row, containing sample size and coef estimates
}
\description{
Simualtes data from a simple linear regression model with user-defined
sample size and regression coefficients. Fits a SLR to the simulated
data and returns regression coefficients.
}
