% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nruns.R
\name{simulate_nruns}
\alias{simulate_nruns}
\title{Repeat simulations}
\usage{
simulate_nruns(n_runs = 100, n, beta0 = 2, beta1 = 3)
}
\arguments{
\item{n}{sample size}

\item{beta0}{intercept}

\item{beta1}{slope}
}
\value{
tibble with `n_runs` rows
}
\description{
Repeat simulations
}
