% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test Function}
\usage{
test_function(x)
}
\arguments{
\item{x}{a sample parameter}
}
\value{
the input parameter duplicated 5 times
}
\description{
Test Function
}
