% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bern_mean.R
\name{sim_bern_mean}
\alias{sim_bern_mean}
\title{Function that Simulates}
\usage{
sim_bern_mean(n_samp, prob)
}
\arguments{
\item{prob}{}
}
\description{
from a different data generating mechanism and edit simulate_nrep to
work with both simulation functions
}
