% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bern_mean.R
\name{sim_bern_mean}
\alias{sim_bern_mean}
\title{Simulate from a Bernoulli distribution}
\usage{
sim_bern_mean(n, prob)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{prob}{probability of success on each trial.}
}
\description{
The sim_bern_mean function has the same general structure as sim_regression()
but simulates data from a Bernoulli distribution and returns the sample average.
}
