% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bern_mean.R
\name{sim_bern_mean}
\alias{sim_bern_mean}
\title{Simuate Bernoulli Data}
\usage{
sim_bern_mean(n, prob)
}
\arguments{
\item{n}{sample size}

\item{prob}{probability of success}
}
\value{
tbl_df with sample size and sample average
}
\description{
Simuate Bernoulli Data
}
