% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nruns.R
\name{simulate_nruns}
\alias{simulate_nruns}
\title{Repeat simulations}
\usage{
simulate_nruns(n_runs, sim_func, ...)
}
\arguments{
\item{n_runs}{number of runs}

\item{sim_func}{the simulation function to repeat}

\item{...}{additional arguments}
}
\value{

}
\description{
Repeat simulations
}
